Add a bosh-lite sample manifest
Provide an example manifest for this release ready to be deployed on
bosh-lite, for development and testing.

This manifest is split for readability, and spruce is required to render it.

The manifest includes a group of jobs configured to start the services
that this release wants to monitor. It includes all the boilerplate
required to start the CF services.

Most of the configuration of the boilerplate is the minimum to make the
services deploy, start and run so that can be monitored.
The values are not right or relevant, and had been extracted from other
examples. The services are not fully functional without a proper installation.

This manifest allows integration with datadog by setting the environment
variable $DATADOG_API_KEY.