Add a HTTP check for consul is leader
Add a HTTP check that will check the local HTTP monitor port to get
the leader API, and compare it with the IP and port of the local consul.

To implement this, we need to check the server port instead of the raft
port, so we add a new property.

With this check, we can create a datadog monitor to check that
instance:consul_is_leader is down or not.