Change consul leader check to be opt-in
I'll set the new option `datadog.consul.leader_checks` in `alphagov/paas-cf`
for all nodes that have `consul.agent.mode: server` (eligible as leader).

We want to run `datadog-consul` in all the places that we run `consul_agent`
because it monitors the availability of the agent in `client` mode, which
affects things like DNS resolution within Cloud Foundry.

We only want the leader check to run on the smaller number of nodes that run
in `server` mode (three, in our case). Running those on client nodes is a
minor waste of resources, but also has the more adverse effect of making it
hard to determine the state of the cluster when a monitor is triggered,
because the DataDog event stream contains reports from every client and
truncates the results:

    No consul cluster servers are repoted as leader
    42 groups are CRITICAL:
    host:uaa.0,instance:consul_is_leader,url:http://localhost:8500/v1/status/leader (last reported 2017-03-20 09:59:32 UTC) message: Content "10.0.16.52:8300" not found in response
    host:access.1,instance:consul_is_leader,url:http://localhost:8500/v1/status/leader (last reported 2017-03-20 09:59:32 UTC) message: Content "10.0.17.9:8300" not found in response
    host:etcd.2,instance:consul_is_leader,url:http://localhost:8500/v1/status/leader (last reported 2017-03-20 09:59:31 UTC) message: Content "10.0.18.19:8300" not found in response
    host:uaa.1,instance:consul_is_leader,url:http://localhost:8500/v1/status/leader (last reported 2017-03-20 09:59:31 UTC) message: Content "10.0.17.51:8300" not found in response
    host:consul.2,instance:consul_is_leader,url:http://localhost:8500/v1/status/leader (last reported 2017-03-20 09:59:31 UTC) message: UP
    host:route_emitter.1,instance:consul_is_leader,url:http://localhost:8500/v1/status/leader (last reported 2017-03-20 09:59:31 UTC) message: Content "10.0.17.43:8300" not found in response
    host:access.0,instance:consul_is_leader,url:http://localhost:8500/v1/status/leader (last reported 2017-03-20 09:59:31 UTC) message: Content "10.0.16.9:8300" not found in response
    host:doppler.1,instance:consul_is_leader,url:http://localhost:8500/v1/status/leader (last reported 2017-03-20 09:59:30 UTC) message: Content "10.0.17.51:8300" not found in response
    …