Split Consul client and server checks
Currently we run the Consul server checks on all VMs, because the
datadog-consul job in this release couples the two together. This
splits them into those for the client and those for the server.

They both have '-agent-' in the name because technically the Consul
agent process runs both.

We also update the sample manifest so it can be tested with bosh-lite,
which should be much quicker than deploying the whole of CloudFoundry
with Datadog enabled.