Add new job to monitor CDN broker
Which queries the new healthcheck endpoint which checks all dependencies
added in alphagov/paas-cdn-broker#3

I'm not adding this to the bosh-lite manifests because we don't have the
BOSH releases for the CDN broker deployed here and I don't think it makes
sense to add them.

I was hoping to set `include_content` so that the monitor contained
information about which healthchecks have failed, but it's not supported by
the version of the agent that we're currently using. I considered leaving it
in so that it works one day but I thought it might be confusing in the
meantime.

- http://docs.datadoghq.com/integrations/httpcheck/