Fix the path to bosh-dns's process.yaml
It looks like the name process.yaml is special to datadog, it won't
work properly if the file is called `bosh_dns_process.yaml`.

It is ok to give the file the same name as all the other process checks
because the config directory lives in `/var/vcap/jobs/datadog-bosh-dns/`
so it's obvious what it's for.